dev/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.9.4
psycopg2
django-widget-tweaks
plotly
numpy
# psycopg2-binary>=2.8
# Django-3.1.3 asgiref-3.3.1 django-widget-tweaks-1.4.8 djangorestframework-3.12.2 numpy-1.19.4 plotly-4.13.0 psycopg2-2.8.6 pytz-2020.4 retrying-1.3.3 six-1.15.0 sqlparse-0.4.1